backend/pyproject.toml
[project]
name = "pereobuyka-backend"
version = "0.1.0"
description = "Ядро системы Переобуйка: бизнес-правила, расписание, прайс, записи, API."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.2",
    "openai>=1.0.0",
    "python-multipart>=0.0.9",
    "sqlglot>=30.6.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "aiosqlite>=0.19.0",
    "testcontainers[postgres]>=4.7.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "SIM"]
ignore = ["B008"]  # FastAPI намеренно использует Depends/Query в дефолтных аргументах

[tool.ruff.lint.per-file-ignores]
"alembic/**" = ["E501"]  # DDL-строки Alembic часто длиннее лимита строки

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
